pyproject.toml
[tool.poetry]
name = "narlun-server"
version = "0.1.0"
description = ""
authors = ["Kris Fris <31852063+krisfris@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = {extras = ["speedups"], version = "^3.7.4"}
colorama = "^0.4.4"
structlog = "^21.2.0"
python-json-logger = "^2.0.2"
isodate = "^0.6.0"
python-dotenv = "^0.19.1"
PyJWT = "^2.2.0"
aiobotocore = "^1.4.2"
asyncpg = "^0.24.0"
uvloop = "^0.16.0"
GeoAlchemy2 = "^0.9.2"
alembic = "^1.7.4"
Werkzeug = "^2.0.2"
pytest-aiohttp = "^0.3.0"
psycopg2-binary = "^2.9.1"
sentry-sdk = "^1.5.3"
Fabric = "^2.7.0"
aiohttp_cors = "^0.7.0"
py-avataaars = "^1.1.2"
email-validator = "^1.2.1"
itsdangerous = "^2.1.2"
aioredis = {extras = ["hiredis"], version = "^2.0.1"}
async-timeout = "^4.0.2"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
pytest-cov = "^3.0.0"
aiohttp-devtools = "^1.0.post0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:aioredis",
	"ignore::RuntimeWarning:aiohttp"
]
